backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2

# Database
motor==3.3.1        # MongoDB async driver
pymongo==4.6.0      # MongoDB sync driver
sqlalchemy==2.0.23  # SQL ORM (if using SQLite instead)

# Utils
python-jose==3.3.0  # JWT tokens for auth
passlib==1.7.4      # Password hashing
python-multipart==0.0.6  # Form data processing
httpx==0.25.1       # Async HTTP client

# AI Integration
g4f==0.1.9.0        # Free GPT API wrapper

# Date and time handling
python-dateutil==2.8.2

# PDF generation
reportlab==4.0.7

# CORS
starlette==0.27.0
